Split off types into their own modules
