+++ packages/liveblocks-client/src/immutable.ts
@@ -1,11 +1,9 @@
 import { AbstractCrdt } from "./AbstractCrdt";
-import type { Json } from "./json";
 import { LiveList } from "./LiveList";
 import { LiveMap } from "./LiveMap";
 import { LiveObject } from "./LiveObject";
 import { LiveRegister } from "./LiveRegister";
-import type { Lson, LsonObject } from "./lson";
-import type { StorageUpdate } from "./types";
+import type { Json, Lson, LsonObject, StorageUpdate } from "./types";
 import { findNonSerializableValue } from "./utils";
 
 function lsonObjectToJson<O extends LsonObject>(
