+++ packages/liveblocks-client/src/index.ts
@@ -1,13 +1,15 @@
 export { createClient } from "./client";
-export type { Json, JsonObject } from "./json";
 export { LiveList } from "./LiveList";
 export { LiveMap } from "./LiveMap";
 export { LiveObject } from "./LiveObject";
-export type { Lson, LsonObject } from "./lson";
 export type {
   BroadcastOptions,
   Client,
   History,
+  Json,
+  JsonObject,
+  Lson,
+  LsonObject,
   Others,
   Presence,
   Room,
