+++ packages/liveblocks-client/test/utils.ts
@@ -1,6 +1,5 @@
 import type { AbstractCrdt } from "../src/AbstractCrdt";
 import { lsonToJson, patchImmutableObject } from "../src/immutable";
-import type { Json, JsonObject } from "../src/json";
 import type {
   ClientMsg,
   Op,
@@ -8,11 +7,16 @@ import type {
   ServerMsg,
 } from "../src/live";
 import { ClientMsgCode, CrdtType, ServerMsgCode } from "../src/live";
-import type { LsonObject, ToJson } from "../src/lson";
 import { makePosition } from "../src/position";
 import type { Effects, Machine } from "../src/room";
 import { defaultState, makeStateMachine } from "../src/room";
-import type { Authentication } from "../src/types";
+import type {
+  Authentication,
+  Json,
+  JsonObject,
+  LsonObject,
+  ToJson,
+} from "../src/types";
 import { remove } from "../src/utils";
 
 /**
